// frontend/lib/api/generated/delegateAttrConfigResponse.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */
import type { DelegateAttrConfigResponseAttrType } from "./delegateAttrConfigResponseAttrType";

/**
 * 属性配置响应
 */
export interface DelegateAttrConfigResponse {
  id: string;
  attrKey: string;
  attrLabel: string;
  attrType: DelegateAttrConfigResponseAttrType;
  sortOrder: number;
  enabled: boolean;
  visible: boolean;
}