// frontend/lib/api/generated/delegateAttrRecordResponse.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */
import type { DelegateAttrRecordResponseValues } from "./delegateAttrRecordResponseValues";

/**
 * 代表属性记录响应
 */
export interface DelegateAttrRecordResponse {
  recordId: string;
  delegateId: string;
  delegateName: string;
  updatedAt: string;
  values: DelegateAttrRecordResponseValues;
}