// frontend/lib/api/generated/delegateAttrTypedValueResponse.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */
import type { DelegateAttrTypedValueResponseAttrType } from "./delegateAttrTypedValueResponseAttrType";

/**
 * 聚合属性值
 */
export interface DelegateAttrTypedValueResponse {
  attrType: DelegateAttrTypedValueResponseAttrType;
  textValue?: string;
  numberValue?: number;
}