// frontend/lib/api/generated/delegateAttrValueInput.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * AsyaMun.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface DelegateAttrValueInput {
  attrKey: string;
  /** @nullable */
  textValue: string | null;
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
   */
  numberValue: number | string | null;
}